+++ src/Shared/Router.fs
@@ -8,17 +8,26 @@ type ArticleRoute =
     | ArticlesList
     | Article of string
 
+type SessionRoute =
+    | Settings
+    | NewPost
+    | Logout
+
 type Route =
     | Login
     | Register
     | Article of ArticleRoute
+    | SessionRoute of SessionRoute
 
 let pageParser: Parser<Route -> Route, Route> =
     oneOf
         [ map (ArticleRoute.Article >> Article) (s "article" </> str)
           map (ArticlesList |> Article) top
           map (Login) (s "login")
-          map (Register) (s "register") ]
+          map (Register) (s "register")
+          map (Settings |> SessionRoute) (s "settings")
+          map (NewPost |> SessionRoute) (s "editor")
+          map (Logout |> SessionRoute) (s "logout") ]
 
 let toHash route =
     match route with
@@ -26,6 +35,9 @@ let toHash route =
     | Article(ArticleRoute.Article slug) -> sprintf "article/%s" slug
     | Login -> "login"
     | Register -> "register"
+    | SessionRoute Settings -> "settings"
+    | SessionRoute NewPost -> "editor"
+    | SessionRoute Logout -> "logout"
     |> (fun r -> sprintf "#/%s" r)
 
 let href = toHash >> Href